Move handle_exception to a module to be reused
