+++ lib/kubeclient.rb
@@ -3,6 +3,7 @@ require 'json'
 require 'rest-client'
 require 'kubeclient/entity_list'
 require 'kubeclient/kube_exception'
+require 'kubeclient/error_handling'
 require 'kubeclient/watch_notice'
 require 'kubeclient/watch_stream'
 require 'kubeclient/common'
